requirements.txt
scispacy==0.5.3
negspacy==1.0.3
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.1/en_ner_bc5cdr_md-0.5.1.tar.gz
Flask==2.0.0
ollama==0.1.9